% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibb.R
\name{fib_code}
\alias{fib_code}
\title{Fib Code Wrapper Function}
\usage{
fib_code(
  prompt,
  history = NULL,
  stream = TRUE,
  model = Sys.getenv("FIBBR_MODEL")
)
}
\arguments{
\item{prompt}{A character string representing the prompt for the R script.}

\item{history}{(Optional) A list of previous interactions to be passed to the \code{fib()} function.}

\item{stream}{(Optional) A logical value indicating whether to stream the output. Default is TRUE.}

\item{model}{(Optional) A character string specifying the model to be used. Define via options, e.g. options(fib_model = "gpt-3.5-turbo").}
}
\value{
The result of the \code{fib()} function call.
}
\description{
This function is a wrapper for the \code{fib()} function, which takes a prompt and additional
arguments to generate an R script based on the given prompt. It also inserts a new line
at the end using the \code{rstudioapi::insertText()} function.
}
\examples{
fib_code("Calculate the sum of 1 to 10")
fib_code("Create a function to find the factorial of a number")
}
