% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibb.R
\name{fib_roxygen}
\alias{fib_roxygen}
\title{Generate an Roxygen Docstring}
\usage{
fib_roxygen(
  prompt,
  history = NULL,
  stream = TRUE,
  model = Sys.getenv("FIBBR_MODEL")
)
}
\arguments{
\item{prompt}{A character string representing the code snippet for which the docstring is to be generated.}

\item{history}{(Optional) A list of previous interactions to be passed to the \code{fib()} function.}

\item{stream}{(Optional) A logical value indicating whether to stream the output. Default is TRUE.}

\item{model}{(Optional) A character string specifying the model to be used. Default is the value set in options("fib_model")[\link{1}].}
}
\value{
A character string containing the generated roxygen docstring.
}
\description{
This function generates a roxygen docstring for a given code snippet without repeating the code.
It utilizes the \code{fib()} function to create the docstring based on the provided prompt and other parameters.
}
\examples{
\dontrun{
  fib_roxygen("example_function = function(x) { return(x + 1) }")
}
}
