% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibb.R
\name{fib_set_token}
\alias{fib_set_token}
\title{Set the API token for the FibbR package}
\usage{
fib_set_token()
}
\value{
NULL, invisibly.
}
\description{
This function sets the API token for the FibbR package using the keyring package.
It prompts the user to enter their Open AI API token and stores it securely in the system keyring.
The token is associated with the "fibbr_openai" service and the current user's username.
}
\examples{
\dontrun{
  fib_set_token()
}
}
