% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibb.R
\name{process_stream_rstudio}
\alias{process_stream_rstudio}
\title{Process Stream RStudio}
\usage{
process_stream_rstudio(data)
}
\description{
This function processes the input data stream and inserts the resulting text into the RStudio console.
}
