% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibb.R
\name{fib}
\alias{fib}
\title{Fib Function}
\usage{
fib(
  prompt,
  model = Sys.getenv("FIBBR_MODEL"),
  system_start_prompt = "You are helpful, but consise.",
  history = NULL,
  stream = FALSE,
  logit_bias = NULL,
  temperature = 0
)
}
\arguments{
\item{prompt}{A character string representing the user's input prompt.}

\item{model}{A character string representing the model to be used.}

\item{system_start_prompt}{A character string for the system's starting prompt (default: "You want to fib.").}

\item{history}{A list of previous interactions (default: NULL).}

\item{stream}{A logical value indicating whether to stream the response (default: TRUE).}
}
\value{
An invisible TRUE value upon successful execution.
}
\description{
This function sends a request to a specified model with a given prompt and optional parameters.
It returns a nested list with the response history, including the latest response.
}
\examples{
\dontrun{
fib("What is the meaning of life?", "gpt-3.5-turbo")
}
}
